demos/use_cases/newsletter_is_model_choice_the_free_lunch_in_ai/pyproject.toml
[tool.poetry]
name = "model-choice-newsletter-code-snippets"
version = "0.1.0"
description = "Benchmarking model alias routing with Arch Gateway."
authors = ["Your Name <your@email.com>"]
license = "Apache 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13.3"
pydantic = "^2.0"
openai = "^1.0"
pyyaml = "^6.0"
archgw ="^0.3.12"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
